requirements.txt
# Email RAG Pipeline Dependencies

# Web Framework
fastapi>=0.109.0
uvicorn>=0.27.0
sse-starlette>=1.8.2

# Database
psycopg2-binary>=2.9.9
pgvector>=0.2.4

# LangChain Core
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-text-splitters>=0.3.0
langchain-postgres>=0.0.6
langchain-core>=0.3.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0

# For evaluation
numpy>=1.24.0
